rename: 공유 링크 생성 함수
